Allow pre-processing of configuration variables
When Get-ConfigurationVariable is called with -Resolved, environment
settings are merged in with application settings and the original context
of the source variable is lost.  There are times when it is important to
know, in order to extend the configuration mechanism, the detail of the
original variable (name, scope, etc.).

The new -ProcessVariable parameter allows a caller to process variables
prior to resolving.