Support execution of scripts in phases
Powerdeploy only supports execution of scripts before and after the install
phase.  It now supports three phases.  Prepare -> Install -> Configure.

Prepare is intended for laying down the platform as necessary.
Install is installation of the application being deployed.
Configure is activity that must happen after the application is installed.

Using the phases is not necessary, but it allows extensions to
deterministically plug into the deployment.